Add messagebox Import, Export, Delete config

+++ page/logic/option.py
@@ -1,3 +1,4 @@
+from tkinter import messagebox
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING: from page.option import option
@@ -5,16 +6,16 @@ if TYPE_CHECKING: from page.option import option
 def export_conf(self: "option"):
     try:
         self.config.exportConfig()
-        self.tm.start("confExport")
-    except:
-        print('export')
+        messagebox.showinfo("Export", "Config is exported")
+    except Exception as e:
+        self.log.error(e)
 
 def import_conf(self: "option"):
     try:
         self.config.importConfig()
-        self.tm.start("confImport")
-    except:
-        print('import')
+        messagebox.showinfo("Import", "Config is imported")
+    except Exception as e:
+        self.log.error(e)
 
 def fixLang(self: "option", event):
     lc = self.langs.getLocaleShort(self.combox_option_lang.get())
@@ -27,5 +28,9 @@ def fixLang(self: "option", event):
 
 def delete_conf(self: "option"):
 
-    self.config.delete()
-    self.tm.start("confDelete")
+    try:
+        if messagebox.askyesno("Delete Config", "Are you sure you want to delete the config ?"):
+            self.config.delete()
+            messagebox.showinfo("Delete Config", "The 'config.json' file has been deleted")
+    except Exception as e:
+        self.log.error(e)
